+++ docs/changelog.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented below.
 
+#### geospatial.analyze.color
+
+* v0.1dev: **plantcv.geospatial.analyze.color**(*img, bin_mask, geojson, bins=10, colorspaces="hsv", label=None*)
 
 #### geospatial.analyze.coverage